# helm/usdsearch/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "USD Search and Asset Graph Search APIs",
...
    "version": "1.4.0"
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "paths": {
...
    "/dependency_graph/flat": {
      "get": {
        "tags": [
          "AGS Asset Graph"
        ],
        "summary": "Get Dependencies Flat",
        "description": "Get a flat list of all files an asset depends on \u2014 simpler alternative to the graph endpoint.\n\nUse this endpoint when you just need the list of referenced files without graph structure\n(e.g., \"list all textures used by this model\"). Returns Asset objects with URLs.\nFor graph structure with edges, use GET /dependency_graph/graph instead.",
        "operationId": "get_dependencies_flat_dependency_graph_flat_get",
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "HTTPBasic": []
          },
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "root_node_url",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "URL of the asset",
              "title": "Root Node Url"
            },
            "description": "URL of the asset"
          },
          {
            "name": "max_level",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "exclusiveMinimum": 0,
              "description": "Max level of dependency tree traversal (by default unlimited)",
              "title": "Max Level"
            },
            "description": "Max level of dependency tree traversal (by default unlimited)"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "exclusiveMinimum": 0,
              "description": "Page size",
              "default": 1000,
              "title": "Limit"
            },
            "description": "Page size"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Asset"
                  },
                  "title": "Response Get Dependencies Flat Dependency Graph Flat Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/download/asset": {
      "get": {
        "tags": [
          "Download"
        ],
        "summary": "Download an asset and its dependencies as a ZIP archive",
        "description": "Fetch a USD asset together with all of its transitive dependency files (sublayers, referenced USDs, textures) and return them as a single ZIP. Files are laid out relative to a shared root directory so the scene's relative references resolve after extraction. A `manifest.json` describing the bundle (root url, included files, any per-file errors) is added at the archive root. Works for every storage backend (S3, Nucleus, local-fs). If the asset has no recorded dependencies (or the asset graph is unavailable) a single-file archive is returned.",
        "operationId": "download_asset_get",
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "HTTPBasic": []
          },
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "asset_url",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "The complete storage URL of the asset to download (e.g. s3://bucket/path/scene.usd).",
              "title": "Asset Url"
            },
            "description": "The complete storage URL of the asset to download (e.g. s3://bucket/path/scene.usd)."
          },
          {
            "name": "manifest_only",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "If true, do not build the ZIP \u2014 return JSON describing the bundle (root url, file count, total size in bytes, per-file sizes, errors). Lets a UI show the bundle size before the user commits to the download.",
              "default": false,
              "title": "Manifest Only"
            },
            "description": "If true, do not build the ZIP \u2014 return JSON describing the bundle (root url, file count, total size in bytes, per-file sizes, errors). Lets a UI show the bundle size before the user commits to the download."
          }
        ],
        "responses": {
          "200": {
            "description": "ZIP archive containing the asset and its dependencies",
            "content": {
              "application/zip": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "403": {
            "description": "Access Denied",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Access denied"
                }
              }
            }
          },
          "404": {
            "description": "Asset Not Found",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Asset 's3://bucket/asset.usd' not found"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Asset": {
        "properties": {
          "url": {
            "type": "string",
            "title": "Url",
            "description": "Full URL of the asset on the storage backend (e.g., omniverse://server/path/file.usd or s3://bucket/key)"
          },
          "deleted": {
            "type": "boolean",
            "title": "Deleted",
            "description": "Whether this asset has been marked as deleted from the storage backend",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "url"
        ],
        "title": "Asset",
        "description": "A file node in the asset dependency graph.\n\nRepresents any file tracked by the Asset Graph Service (USD scenes, textures, materials, etc.).\nThe `url` uniquely identifies the asset on the storage backend."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      },
      "HTTPBasic": {
        "type": "http",
        "scheme": "basic"
      },
      "APIKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key"
      }
    }
  },
...
}